backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pandas
numpy
pydantic
transformers
torch
scipy
langgraph
langchain-google-genai
tqdm
